consolidate behavior tests
